make export patterns more consistent
